# applications/rpg_table_helper/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Main API v1.0",
    "version": "v1.0"
  },
  "paths": {
...
    "/Notes/getdocument/{notedocumentid}": {
      "get": {
        "tags": [
          "Notes"
        ],
        "summary": "Returns a single document.",
        "description": "You must be within the correct campagne and the document has to be your own or shared with you",
        "parameters": [{
          "name": "notedocumentid",
          "in": "path",
          "description": "The id of the desired campagne",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "The document",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDocumentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDocumentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDocumentDto"
                }
              }
            }
          },
          "400": {
            "description": "If there was an error retrieving the document",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If you are not logged in or you are not allowed to see this document",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/Notes/deletedocument/{notedocumentid}": {
      "delete": {
        "tags": [
          "Notes"
        ],
        "summary": "Deletes a single document.",
        "description": "You must be within the correct campagne and the document has to be your own",
        "parameters": [{
          "name": "notedocumentid",
          "in": "path",
          "description": "The id of the desired campagne",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "The document",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDocumentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDocumentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDocumentDto"
                }
              }
            }
          },
          "400": {
            "description": "If there was an error retrieving the document",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If you are not logged in or you are not allowed to see this document",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/Notes/createdocument": {
      "post": {
        "tags": [
          "Notes"
        ],
        "summary": "Creates a single document.",
        "description": "You must be within the correct campagne",
        "requestBody": {
          "description": "The new document",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/NoteDocumentDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteDocumentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteDocumentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NoteDocumentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The id of the new document",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDocumentIdentifier"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDocumentIdentifier"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDocumentIdentifier"
                }
              }
            }
          },
          "400": {
            "description": "If there was an error retrieving the document",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If you are not logged in or you are not allowed to see this document",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/Notes/createtextblock/{notedocumentid}": {
      "post": {
        "tags": [
          "Notes"
        ],
        "summary": "Creates a single text block for a given document.",
        "description": "You must be the owner of the document",
        "parameters": [{
          "name": "notedocumentid",
          "in": "path",
          "description": "The document id where this block will be assigned",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "requestBody": {
          "description": "The new textblock",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/TextBlock"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextBlock"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TextBlock"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TextBlock"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The id of the new document",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NoteBlockModelBaseIdentifier"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteBlockModelBaseIdentifier"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteBlockModelBaseIdentifier"
                }
              }
            }
          },
          "400": {
            "description": "If there was an error retrieving the document",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If you are not logged in or you are not allowed to see this document",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HttpValidationProblemDetails": {
        "allOf": [{
            "$ref": "#/components/schemas/ProblemDetails"
          },
          {
            "type": "object",
            "properties": {
              "errors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "nullable": true
              }
            },
            "additionalProperties": {}
          }
        ]
      },
...
      "NoteBlockModelBaseIdentifier": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "NoteDocumentDto": {
        "required": [
          "createdForCampagneId",
          "groupName",
          "imageBlocks",
          "textBlocks",
          "title"
        ],
        "type": "object",
        "properties": {
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/NoteDocumentIdentifier"
          },
          "groupName": {
            "minLength": 1,
            "type": "string"
          },
          "creatingUserId": {
            "$ref": "#/components/schemas/UserIdentifier"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "createdForCampagneId": {
            "$ref": "#/components/schemas/CampagneIdentifier"
          },
          "imageBlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageBlock"
            }
          },
          "textBlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TextBlock"
            }
          }
        },
        "additionalProperties": false
      },
      "NoteDocumentIdentifier": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "TextBlock": {
        "allOf": [{
            "$ref": "#/components/schemas/NoteBlockModelBase"
          },
          {
            "type": "object",
            "properties": {
              "markdownText": {
                "type": "string",
                "nullable": true
              }
            },
            "additionalProperties": false
          }
        ]
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 1safsfsdfdfd\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [{
    "Bearer": []
  }]
}